services/auth-service/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
httpx==0.27.0
pyjwt==2.8.0
pydantic==2.7.1
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-httpx==0.45b0
